Salvando localizacao e alterando movendo o metodo para a controller do dev
