+++ src/models/Dev.js
@@ -1,11 +1,16 @@
-const mongoose = require('mongoose');
+const mongoose    = require('mongoose');
+const PointSchema = require('./utils/PointSchema');
 
 const DevSchema = new mongoose.Schema({
     name: String,
     github_username: String,
     bio: String,
     avatar_url: String,
-    techs: [String]
+    techs: [String],
+    location: {
+        type: PointSchema,
+        index: '2dsphere'
+    }
 
 });
