+++ src/models/utils/PointSchema.js
@@ -1,0 +1,15 @@
+const mongoose = require('mongoose');
+
+const PointSchema = new mongoose.Schema({
+    type: {
+        type: String,
+        enum: ['Point'],
+        required: true
+    },
+    coordinates: {
+        type: [Number],
+        required: true
+    }
+});
+
+module.exports = PointSchema;
