+++ src/routes.js
@@ -1,28 +1,9 @@
 const { Router } = require('express');
-const axios      = require('axios');
-const Dev        = require('./models/Dev');
+const DevController = require('./controllers/DevController')
 
 const routes = Router();
 
 
-routes.post('/devs', async (request, response) => {
-    const { github_username, techs } = request.body;
-
-    const apiResponse = await axios.get(`https://api.github.com/users/${github_username}`);
-
-    const { name = login, avatar_url, bio} = apiResponse.data;
-
-    const techsArray = techs.split(',').map(tech => tech.trim());
-
-    const dev = await Dev.create({
-        github_username,
-        name,
-        avatar_url,
-        bio,
-        techs: techsArray
-    });
-
-    return response.json(dev);
-});
+routes.post('/devs', DevController.store);
 
 module.exports = routes;
